Refactor project
I felt that the original iteration of the project
was entirely unmaintainable if I wanted to expand
the feature set at any point. Because of this, I
have refactored a decent chunk of the codebase,
and reorganised the package. With this structure,
I should be able to add features without much
trouble.

== CHANGELOG ==

* Created a new _types.py file to encompass all
new classes used by the package.

* Created a new util.py file to include any
useful functions that are used throughout the
package.

* Refactored reader.py to use the new ColorSwatch
class and ColorSpace enum.

* Other small miscellaneous changes.